refactor: move to domain managers
