+++ src/operators/gitea/gitea.ts
@@ -1,46 +1,30 @@
 /* eslint-disable no-console */
-import { CoreV1Api, Exec, KubeConfig, KubernetesObject, V1Status } from '@kubernetes/client-node'
+import { KubeConfig } from '@kubernetes/client-node'
 import Operator, { ResourceEvent, ResourceEventType } from '@linode/apl-k8s-operator'
 import {
   AdminApi,
   Configuration,
-  CreateHookOption,
-  CreateHookOptionTypeEnum,
   CreateRepoOption,
-  CreateTeamOption,
-  CreateTeamOptionPermissionEnum,
-  EditHookOption,
-  EditRepoOption,
   MiscellaneousApi,
   Organization,
   OrganizationApi,
   Repository,
   RepositoryApi,
-  ResponseError,
-  Team,
-  User,
 } from '@linode/gitea-client-fetch'
 import retry from 'async-retry'
-import { generate as generatePassword } from 'generate-password'
 import { isEmpty, keys } from 'lodash'
-import stream from 'stream'
-import { getRepoNameFromUrl, setServiceAccountSecret } from '../../gitea-utils'
-import { getTektonPipeline } from '../../k8s'
 import { getSanitizedErrorMessage } from '../../utils'
-import { orgName, otomiChartsRepoName, otomiValuesRepoName, teamNameOwners, teamNameViewer } from '../common'
+import { orgName, otomiValuesRepoName } from '../common'
 import { giteaEnv } from './lib/env'
-
-// Interfaces
-interface hookInfo {
-  id?: number
-  hasHook: boolean
-}
-
-interface groupMapping {
-  [key: string]: {
-    [teamId: string]: string[]
-  }
-}
+import { createOrganizations } from './lib/managers/gitea-organizations'
+import { errors } from './lib/globals'
+import { createTeams } from './lib/managers/gitea-teams'
+import { createReposAndAddToTeam, upsertRepo } from './lib/managers/gitea-repositories'
+import { createUsers } from './lib/managers/gitea-users'
+import { setGiteaOIDCConfig } from './lib/managers/gitea-oidc'
+import { addTektonHook, createBuildWebHook, deleteBuildWebHook, updateBuildWebHook } from './lib/managers/gitea-webhook'
+import { PipelineKubernetesObject } from './lib/types/webhook'
+import { getTektonPipeline } from './lib/helpers'
 
 interface DependencyState {
   giteaPassword: string | null
@@ -50,32 +34,20 @@ interface DependencyState {
   oidcEndpoint: string | null
   teamNames: string[] | null
 }
-interface Param {
-  name: string
-  value: string
-}
-interface Task {
-  name: string
-  params: Param[]
-}
-
-interface PipelineTemplateObject extends KubernetesObject {
-  spec: {
-    pipelineRef: {
-      name: string
-    }
-  }
-}
-
-export interface PipelineKubernetesObject extends KubernetesObject {
-  spec: {
-    tasks: Task[]
-    resourcetemplates: PipelineTemplateObject[]
-  }
-}
 
 const GITEA_ENDPOINT = `${giteaEnv.GITEA_URL}:${giteaEnv.GITEA_URL_PORT}`
-const env = {
+export type GiteaConfig = {
+  giteaPassword: string
+  hasArgocd: boolean
+  teamConfig: {}
+  teamNames: string[]
+  domainSuffix: string
+  oidcClientId: string
+  oidcClientSecret: string
+  oidcEndpoint: string
+}
+
+const env: GiteaConfig = {
   giteaPassword: '',
   hasArgocd: false,
   teamConfig: {},
@@ -93,23 +65,6 @@ let lastState: DependencyState = {
   oidcEndpoint: null,
   teamNames: null,
 }
-const errors: string[] = []
-
-const readOnlyTeam: CreateTeamOption = {
-  canCreateOrgRepo: false,
-  name: teamNameViewer,
-  includesAllRepositories: false,
-  permission: CreateTeamOptionPermissionEnum.Read,
-  units: ['repo.code'],
-}
-
-const editorTeam: CreateTeamOption = {
-  ...readOnlyTeam,
-  includesAllRepositories: false,
-  permission: CreateTeamOptionPermissionEnum.Write,
-}
-
-const adminTeam: CreateTeamOption = { ...editorTeam, permission: CreateTeamOptionPermissionEnum.Admin }
 
 const kc = new KubeConfig()
 // loadFromCluster when deploying on cluster
@@ -118,18 +73,6 @@ if (process.env.KUBERNETES_SERVICE_HOST && process.env.KUBERNETES_SERVICE_PORT) 
   kc.loadFromCluster()
 } else {
   kc.loadFromDefault()
-}
-
-function isUnprocessableError(error): boolean {
-  return error instanceof ResponseError && error.response.status === 422
-}
-
-function isNotFoundError(error): boolean {
-  return error instanceof ResponseError && error.response.status === 404
-}
-
-function isNotModifiedError(error) {
-  return error instanceof ResponseError && error.response.status === 304
 }
 
 // Callbacks
@@ -232,120 +175,12 @@ async function triggerTemplateCallback(resourceEvent: ResourceEvent): Promise<vo
   } else return
 }
 
-// Exported for testing purposes
-export const addServiceAccountToOrganizations = async (
-  organizationApi: OrganizationApi,
-  serviceAcountName: string,
-  organizations: Organization[],
-) => {
-  const organization = organizations.find((org) => serviceAcountName === `organization-${org.name}`)
-  let teams: Team[]
-  try {
-    console.info(`Getting teams from organization: ${organization?.name}`)
-    teams = await organizationApi.orgListTeams({ org: organization!.name! })
-  } catch (e) {
-    errors.push(`Error getting teams from organization ${organization?.name}: ${e}`)
-    return
-  }
-  const ownerTeam = teams.find((team) => team.name === 'Owners')
-  let members: User[]
-  try {
-    console.info(`Getting members from Owners team in ${organization?.name}`)
-    members = await organizationApi.orgListTeamMembers({ id: ownerTeam!.id! })
-  } catch (e) {
-    errors.push(`Error getting members from Owners team in ${organization?.name}: ${e}`)
-    return
-  }
-  if (isEmpty(members)) return
-  const exists = members.some((member) => member.login === serviceAcountName)
-  if (exists) return
-  try {
-    console.info(`Adding user to organization Owners team in ${organization?.name}`)
-    await organizationApi.orgAddTeamMember({ id: ownerTeam!.id!, username: serviceAcountName })
-  } catch (e) {
-    errors.push(`Error adding user to organization Owners team in ${organization?.name}: ${e}`)
-  }
-}
-
-// Exported for testing purposes
-export const editServiceAccount = async (adminApi: AdminApi, loginName: string, password: string) => {
-  const editUserOption = {
-    sourceId: 0,
-    loginName,
-    password,
-  }
-  try {
-    console.info(`Editing user: ${loginName} with new password`)
-    await adminApi.adminEditUser({ username: loginName, body: editUserOption })
-  } catch (e) {
-    errors.push(`Error editing user ${loginName}: ${e}`)
-  }
-}
-
-// Exported for testing purposes
-export const createServiceAccounts = async (
-  adminApi: AdminApi,
-  organizations: Organization[],
-  orgApi: OrganizationApi,
-) => {
-  let users: User[]
-  try {
-    console.info('Getting all users')
-    users = await adminApi.adminSearchUsers()
-  } catch (e) {
-    errors.push(`Error getting all users: ${e}`)
-    return
-  }
-  const filteredOrganizations = organizations.filter((org) => org.name !== 'otomi')
-  await Promise.all(
-    filteredOrganizations.map(async (organization) => {
-      const serviceAccountSecretName = 'gitea-credentials'
-      const exists = users.some((user) => user.login === `organization-${organization.name}`)
-      const password = generatePassword({
-        length: 16,
-        numbers: true,
-        symbols: true,
-        lowercase: true,
-        uppercase: true,
-        exclude: String(':,;"/=|%\\\''),
-      })
-      const giteaURL = `https://gitea.${env.domainSuffix}`
-      const serviceAccount = `organization-${organization.name}`
-
-      if (!exists) {
-        const organizationEmail = `${organization.name}@mail.com`
-        const createUserOption = {
-          email: organizationEmail,
-          password,
-          username: serviceAccount,
-          loginName: serviceAccount,
-          fullName: serviceAccount,
-          restricted: false,
-          mustChangePassword: false,
-          repoAdminChangeTeamAccess: true,
-        }
-        try {
-          console.info(`Creating user: ${serviceAccount}`)
-          await adminApi.adminCreateUser({ body: createUserOption })
-        } catch (e) {
-          errors.push(`Error creating user ${serviceAccount}: ${e}`)
-        }
-      } else {
-        await editServiceAccount(adminApi, serviceAccount, password)
-      }
-
-      await setServiceAccountSecret(serviceAccountSecretName, serviceAccount, organization.name!, password, giteaURL)
-      await addServiceAccountToOrganizations(orgApi, serviceAccount, filteredOrganizations)
-    }),
-  )
-}
-
 const createSetGiteaOIDCConfig = (() => {
   let intervalId: any = null
   return function runSetGiteaOIDCConfig() {
     if (intervalId === null) {
       intervalId = setInterval(() => {
-        setGiteaOIDCConfig()
+        setGiteaOIDCConfig(env, kc)
           .catch((error) => {
             console.error('Error occurred during setGiteaOIDCConfig execution:', error)
           })
@@ -466,7 +301,7 @@ async function checkAndExecute() {
     currentState.oidcEndpoint !== lastState.oidcEndpoint ||
     currentState.teamNames !== lastState.teamNames
   ) {
-    await setGiteaOIDCConfig(true)
+    await setGiteaOIDCConfig(env, kc, true)
   }
 
   // Update last known state
@@ -483,336 +318,6 @@ async function runSetupGitea() {
     await new Promise((resolve) => setTimeout(resolve, 30000))
     console.debug('Retrying to setup gitea')
     await runSetupGitea()
-  }
-}
-
-// Exported for testing purposes
-export async function upsertOrganization(
-  orgApi: OrganizationApi,
-  existingOrganizations: Organization[],
-  organizationName: string,
-): Promise<Organization> {
-  const prefixedOrgName = !organizationName.includes('otomi') ? `team-${organizationName}` : organizationName
-  const orgOption = {
-    username: prefixedOrgName,
-    fullName: prefixedOrgName,
-    repoAdminChangeTeamAccess: true,
-  }
-  const existingOrg = existingOrganizations.find((organization) => organization.name === prefixedOrgName)
-  if (isEmpty(existingOrg)) {
-    try {
-      console.info(`Creating organization "${orgOption.fullName}"`)
-      return await orgApi.orgCreate({ organization: orgOption })
-    } catch (e) {
-      if (!isUnprocessableError(e)) {
-        errors.push(`Error creating organization "${orgOption.fullName}": ${e}`)
-      }
-      throw e
-    }
-  }
-
-  try {
-    console.info(`Updating organization "${orgOption.fullName}"`)
-    return await orgApi.orgEdit({ org: prefixedOrgName, body: orgOption })
-  } catch (e) {
-    if (!isUnprocessableError(e)) {
-      errors.push(`Error updating organization "${orgOption.fullName}": ${e}`)
-    }
-    throw e
-  }
-}
-
-// Setup Gitea Functions
-async function upsertTeam(
-  orgApi: OrganizationApi,
-  organizationName: string,
-  teamOption: CreateTeamOption,
-): Promise<void> {
-  const getErrors: string[] = []
-  let existingTeams: Team[]
-  try {
-    console.info(`Getting all teams in organization "${organizationName}"`)
-    existingTeams = await orgApi.orgListTeams({ org: organizationName })
-  } catch (e) {
-    getErrors.push(`Error getting all teams in organization "${organizationName}": ${e}`)
-    console.error('Errors when getting teams.', getErrors)
-    return
-  }
-  const existingTeam = existingTeams?.find((team) => team.name === teamOption.name)
-  if (existingTeam === undefined) {
-    try {
-      console.info(`Creating team "${teamOption.name}" in organization "${organizationName}"`)
-      await orgApi.orgCreateTeam({ org: organizationName, body: teamOption })
-    } catch (e) {
-      if (!isUnprocessableError(e)) {
-        errors.push(`Error creating team "${teamOption.name}" in organization "${organizationName}": ${e}`)
-      }
-    }
-  } else {
-    try {
-      console.info(`Updating team "${teamOption.name}" in organization "${organizationName}"`)
-      await orgApi.orgEditTeam({ id: existingTeam.id!, body: teamOption })
-    } catch (e) {
-      if (!isUnprocessableError(e)) {
-        errors.push(`Error updating team "${teamOption.name}" in organization "${organizationName}": ${e}`)
-      }
-    }
-  }
-}
-
-async function upsertRepo(
-  existingReposInOrg: Repository[] = [],
-  orgApi: OrganizationApi,
-  repoApi: RepositoryApi,
-  repoOption: CreateRepoOption | EditRepoOption,
-  teamName?: string,
-): Promise<void> {
-  const repoName = repoOption.name!
-  const existingRepo = existingReposInOrg.find((repository) => repository.name === repoName)
-  let addTeam = false
-  if (isEmpty(existingRepo)) {
-    // organization repo create
-    console.info(`Creating repo "${repoName}" in organization "${orgName}"`)
-    await orgApi.createOrgRepo({ org: orgName, body: repoOption as CreateRepoOption })
-    addTeam = true
-  } else {
-    // repo update
-    console.info(`Updating repo "${repoName}" in organization "${orgName}"`)
-    await repoApi.repoEdit({ owner: orgName, repo: repoName, body: repoOption as EditRepoOption })
-    if (teamName) {
-      console.info(`Checking if repo "${repoName}" is assigned to team "${teamName}"`)
-      try {
-        await repoApi.repoCheckTeam({ owner: orgName, repo: repoName, team: teamName })
-      } catch (error) {
-        if (isNotFoundError(error)) {
-          addTeam = true
-        } else {
-          throw error
-        }
-      }
-    }
-  }
-  if (addTeam && teamName) {
-    console.info(`Adding repo "${repoName}" to team "${teamName}"`)
-    await repoApi.repoAddTeam({ owner: orgName, repo: repoName, team: teamName })
-  }
-}
-
-async function createOrgsAndTeams(
-  orgApi: OrganizationApi,
-  existingOrganizations: Organization[],
-  organizationNames: string[],
-  teamIds: string[],
-): Promise<Organization[]> {
-  await Promise.all(
-    organizationNames.map(async (organizationName) => {
-      const organization = await upsertOrganization(orgApi, existingOrganizations, organizationName)
-      if (existingOrganizations.find((org) => org.id === organization.id)) return
-      existingOrganizations.push(organization)
-    }),
-  )
-  await Promise.all(
-    teamIds.map((teamId) => {
-      const name = `team-${teamId}`
-      return upsertTeam(orgApi, orgName, { ...adminTeam, name })
-    }),
-  )
-  // create organization wide viewer team for otomi role "team-viewer"
-  await upsertTeam(orgApi, orgName, readOnlyTeam)
-  return existingOrganizations
-}
-
-async function hasSpecificHook(repoApi: RepositoryApi, hookToFind: string): Promise<hookInfo> {
-  let hooks: any[]
-  try {
-    console.debug(`Getting hooks in repo "otomi/values"`)
-    hooks = await repoApi.repoListHooks({ owner: orgName, repo: 'values' })
-  } catch (e) {
-    if (isNotFoundError(e)) {
-      errors.push(`Error getting hooks in repo "otomi/values": ${e}`)
-    }
-    console.debug(`No hooks were found in repo "otomi/values"`)
-    return { hasHook: false }
-  }
-  if (!hooks) {
-    console.debug(`No hooks were found in repo "otomi/values"`)
-    return { hasHook: false }
-  }
-
-  const foundHook = hooks.find((hook) => {
-    return hook.config && hook.config.url.includes(hookToFind)
-  })
-  if (foundHook) {
-    console.debug(`Hook (${hookToFind}) exists in repo "values"`)
-    return { id: foundHook.id, hasHook: true }
-  }
-  console.debug(`Hook (${hookToFind}) not found in repo "values"`)
-  return { hasHook: false }
-}
-
-async function addTektonHook(repoApi: RepositoryApi): Promise<void> {
-  console.debug('Check for Tekton hook')
-  const clusterIP = 'http://el-otomi-tekton-listener.otomi-pipelines.svc.cluster.local:8080'
-  const hasTektonHook = await hasSpecificHook(repoApi, 'el-otomi-tekton-listener')
-  if (!hasTektonHook.hasHook) {
-    console.debug('Tekton Hook needs to be created')
-    try {
-      console.debug(`Adding hook "tekton" to repo otomi/values`)
-      await repoApi.repoCreateHook({
-        owner: orgName,
-        repo: 'values',
-        body: {
-          type: CreateHookOptionTypeEnum.Gitea,
-          active: true,
-          config: {
-            url: clusterIP,
-            http_method: 'post',
-            content_type: 'json',
-          },
-          events: ['push'],
-        },
-      })
-    } catch (e) {
-      if (!isNotModifiedError(e)) {
-        errors.push(`Error adding hook "tekton" to repo otomi/values: ${e}`)
-      }
-    }
-  }
-}
-
-async function createReposAndAddToTeam(
-  orgApi: OrganizationApi,
-  repoApi: RepositoryApi,
-  existingRepos: Repository[],
-  repoOption: CreateRepoOption,
-) {
-  // create main organization repo: otomi/values
-  await upsertRepo(existingRepos, orgApi, repoApi, repoOption)
-  // create otomi/charts repo for auto image updates
-  await upsertRepo(existingRepos, orgApi, repoApi, { ...repoOption, name: otomiChartsRepoName })
-
-  // add repo: otomi/values to the team: otomi-viewer
-  const existingValuesRepo = existingRepos.find((repo) => repo.name === otomiValuesRepoName)
-  const existingChartsRepo = existingRepos.find((repo) => repo.name === otomiChartsRepoName)
-  if (!existingValuesRepo) {
-    try {
-      console.info(`Adding repo "${otomiValuesRepoName}" to team "${teamNameViewer}"`)
-      await repoApi.repoAddTeam({ owner: orgName, repo: otomiValuesRepoName, team: teamNameViewer })
-    } catch (e) {
-      if (!isUnprocessableError(e)) {
-        errors.push(`Error adding repo ${otomiValuesRepoName} to team ${teamNameViewer}: ${e}`)
-      }
-    }
-  }
-  if (!existingChartsRepo) {
-    // add repo: otomi/charts to the team: otomi-viewer
-    try {
-      console.info(`Adding repo "${otomiChartsRepoName}" to team "${teamNameViewer}"`)
-      await repoApi.repoAddTeam({ owner: orgName, repo: otomiChartsRepoName, team: teamNameViewer })
-    } catch (e) {
-      if (!isUnprocessableError(e)) {
-        errors.push(`Error adding repo ${otomiChartsRepoName} to team ${teamNameViewer}: ${e}`)
-      }
-    }
-  }
-}
-
-// Logic to create a webhook for repos in an organization
-export async function createBuildWebHook(
-  repoApi: RepositoryApi,
-  teamName: string,
-  buildWorkspace: { buildName: string; repoUrl: string },
-) {
-  try {
-    const repoName = getRepoNameFromUrl(buildWorkspace.repoUrl)!
-
-    // Check to see if a webhook already exists with the same url and push event
-    const webhooks = await repoApi.repoListHooks({ owner: teamName, repo: repoName })
-    let webhookExists
-    if (!isEmpty(webhooks)) {
-      webhookExists = webhooks.find((hook) => {
-        return (
-          hook.config!.url ===
-            `http://el-gitea-webhook-${buildWorkspace.buildName}.${teamName}.svc.cluster.local:8080` &&
-          hook.events?.includes('push')
-        )
-      })
-    }
-
-    if (!isEmpty(webhookExists)) return
-    const createHookOption: CreateHookOption = {
-      active: true,
-      type: CreateHookOptionTypeEnum.Gitea,
-      events: ['push'],
-      config: {
-        content_type: 'json',
-        url: `http://el-gitea-webhook-${buildWorkspace.buildName}.${teamName}.svc.cluster.local:8080`,
-      },
-    }
-    await repoApi.repoCreateHook({ owner: teamName, repo: repoName, body: createHookOption })
-    console.info(`Gitea webhook created for repository: ${repoName} in ${teamName}`)
-  } catch (error) {
-    throw new Error(`Error creating Gitea webhook`)
-  }
-}
-
-// Logic to update a webhook for repos in an organization
-export async function updateBuildWebHook(
-  repoApi: RepositoryApi,
-  teamName: string,
-  buildWorkspace: { buildName: string; repoUrl: string },
-) {
-  try {
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const repoName = getRepoNameFromUrl(buildWorkspace.repoUrl)!
-    const webhooks = await repoApi.repoListHooks({ owner: teamName, repo: repoName })
-
-    if (isEmpty(webhooks)) {
-      console.debug(`No webhooks found for ${repoName} in ${teamName}`)
-      console.debug('Trying to create one instead...')
-      return await createBuildWebHook(repoApi, teamName, buildWorkspace)
-    }
-
-    const editHookOption: EditHookOption = {
-      active: true,
-      events: ['push'],
-      config: {
-        content_type: 'json',
-        url: `http://el-gitea-webhook-${buildWorkspace.buildName}.${teamName}.svc.cluster.local:8080`,
-      },
-    }
-    await Promise.all(
-      webhooks.map(async (webhook) => {
-        await repoApi.repoEditHook({ owner: teamName, repo: repoName, id: webhook.id!, body: editHookOption })
-      }),
-    )
-    console.info(`Gitea webhook updated for repository: ${repoName} in ${teamName}`)
-  } catch (error) {
-    throw new Error('Error updating Gitea webhook')
-  }
-}
-
-// Logic to delete a webhook for repos in a organization
-export async function deleteBuildWebHook(
-  repoApi: RepositoryApi,
-  teamName: string,
-  buildWorkspace: { buildName: string; repoUrl: string },
-) {
-  try {
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const repoName = buildWorkspace.repoUrl.split('/').pop()!
-    const webhooks = await repoApi.repoListHooks({ owner: teamName, repo: repoName })
-
-    if (isEmpty(webhooks)) throw new Error(`No webhooks found for ${repoName} in ${teamName}`)
-
-    await Promise.all(
-      webhooks.map(async (webhook) => {
-        await repoApi.repoDeleteHook({ owner: teamName, repo: repoName, id: webhook.id! })
-      }),
-    )
-    console.info(`Gitea webhook deleted for repository: ${repoName} in ${teamName}`)
-  } catch (error) {
-    throw new Error('Error deleting Gitea webhook')
   }
 }
 
@@ -838,8 +343,9 @@ async function setupGitea() {
     errors.push(`Error getting all organizations: ${e}`)
     return
   }
-  existingOrganizations = await createOrgsAndTeams(orgApi, existingOrganizations, orgNames, teamIds)
-  await createServiceAccounts(adminApi, existingOrganizations, orgApi)
+  existingOrganizations = await createOrganizations(orgApi, existingOrganizations, orgNames)
+  await createTeams(teamIds, orgApi)
+  await createUsers(adminApi, existingOrganizations, orgApi, env.domainSuffix)
   let existingRepos: Repository[]
   try {
     console.info(`Getting all repos in organization "${orgName}"`)
@@ -873,98 +379,5 @@ async function setupGitea() {
     process.exit(1)
   } else {
     console.info('Success! Gitea setup/reconfiguration completed')
-  }
-}
-
-// Set Gitea Functions
-// Exported for testing purposes
-export function buildTeamString(teamNames: any[]): string {
-  const teamObject: groupMapping = { 'platform-admin': { otomi: [teamNameOwners] } }
-  if (teamNames === undefined) return JSON.stringify(teamObject)
-  teamNames.forEach((teamName: string) => {
-    const team = `team-${teamName}`
-    teamObject[team] = {
-      otomi: [teamNameViewer, team],
-      [team]: ['Owners'],
-    }
-  })
-  return JSON.stringify(teamObject)
-}
-
-async function getGiteaPodName(namespace: string): Promise<string | undefined> {
-  const k8sApi = kc.makeApiClient(CoreV1Api)
-  const giteaPods = await k8sApi.listNamespacedPod({
-    namespace,
-    labelSelector: 'app.kubernetes.io/instance=gitea,app.kubernetes.io/name=gitea',
-    limit: 1,
-  })
-  if (giteaPods.items.length === 0) {
-    console.debug('Not ready for setting up OIDC config: Gitea pod not found.')
-    return
-  }
-  return giteaPods.items[0].metadata?.name
-}
-
-async function setGiteaOIDCConfig(update = false) {
-  if (!env.oidcClientId || !env.oidcClientSecret || !env.oidcEndpoint) return
-  const podNamespace = 'gitea'
-  const clientID = env.oidcClientId
-  const clientSecret = env.oidcClientSecret
-  const discoveryURL = `${env.oidcEndpoint}/.well-known/openid-configuration`
-  const teamNamespaceString = buildTeamString(env.teamNames)
-
-  const podName = await getGiteaPodName(podNamespace)
-  if (!podName) {
-    console.debug('Not ready for setting up OIDC config: Name of Gitea pod not found.')
-    return
-  }
-
-  try {
-    // WARNING: Dont enclose the teamNamespaceString in double quotes, this will escape the string incorrectly and breaks OIDC group mapping in gitea
-    const execCommand = [
-      'sh',
-      '-c',
-      `
-      AUTH_ID=$(gitea admin auth list --vertical-bars | grep -E "\\|otomi-idp\\s+\\|" | grep -iE "\\|OAuth2\\s+\\|" | awk -F " " '{print $1}' | tr -d '\\n')
-      if [ -z "$AUTH_ID" ]; then
-        echo "Gitea OIDC config not found. Adding OIDC config for otomi-idp."
-        gitea admin auth add-oauth --name "otomi-idp" --key "${clientID}" --secret "${clientSecret}" --auto-discover-url "${discoveryURL}" --provider "openidConnect" --admin-group "platform-admin" --group-claim-name "groups" --group-team-map '${teamNamespaceString}'
-      elif ${update}; then
-        echo "Gitea OIDC config is different. Updating OIDC config for otomi-idp."
-        gitea admin auth update-oauth --id "$AUTH_ID" --key "${clientID}" --secret "${clientSecret}" --auto-discover-url "${discoveryURL}" --group-team-map '${teamNamespaceString}'
-      else
-        echo "Gitea OIDC config is up to date."
-      fi
-      `,
-    ]
-    const exec = new Exec(kc)
-    const outputStream = new stream.PassThrough()
-    let output = ''
-    outputStream.on('data', (chunk) => {
-      output += chunk.toString()
-    })
-    // Run gitea CLI command to create/update the gitea oauth configuration
-    await exec
-      .exec(
-        podNamespace,
-        podName,
-        'gitea',
-        execCommand,
-        outputStream,
-        process.stderr as stream.Writable,
-        process.stdin as stream.Readable,
-        false,
-        (status: V1Status) => {
-          console.info(output.trim())
-          console.info('Gitea OIDC config status:', status.status)
-        },
-      )
-      .catch((error) => {
-        console.debug('Error occurred during exec:', error)
-        throw error
-      })
-  } catch (error) {
-    console.debug(`Error Gitea OIDC config: ${error.message}`)
-    throw error
   }
 }
