+++ src/operators/gitea/lib/globals.ts
@@ -1,0 +1,1 @@
+export const errors: string[] = []
